Tests pass with extras that are lists and dicts
